LayoutTests/js/dom/proxy-is-not-allowed-in-global-prototype-chain-expected.txt

CONSOLE MESSAGE: TypeError: Proxy is not allowed in the global prototype chain.
onerror saw TypeError: Proxy is not allowed in the global prototype chain.
